requirements.txt
streamlit
pdfplumber
pandas
requests
google-generativeai